Add ability for holes to specify a constant as input.
This allows for near compatibility with phpgolf.org.  Adding some
helpers to do trim and to further restrict the allowed functions is all
that comes to mind to create compatibility.

I'd still like it if the executes could take other types of input.
Instead of 'constants', 'variables' could be nice for some things, and
STDIN could definitely make sense.

createImage is a nice abstraction of that logic, and I like having the
example script fizzbuzz.php running in a container using this.  It makes
it easier to show that things are indeed running correctly and not tied
to a particular server.