Use proc_open to avoid trimming output.
exec returns output as an array of lines, but all newlines are trimmed
off of the lines so it would be impossible to use it when newlines are
important at the end, for instance.  proc_open gets around that and
gives us access to stderr which could be more useful later on.