Factor out image building logic into buildImage.
Easing the ability to layer up images should make it easier to add some
more functionality.  Anytime we want to persist changes made to the
environment the scripts will run in, we need to create a new image.
This handles some of the basic functionality that such scripts could
want to use.