Allow for functionality to be disabled on a single hole.
Some functionality of the language should be allowed in general, but
sometimes needs to be disabled to make a hole competitive in that
language.

This adds a requirement for languages to support a disableFunctionality
hook that should disable anything necessary for the passed
functionality.

Holes can now request for functionality to be disabled.

The ucwords hole given is included as an example of how this
functionality can be used.