Upgrade to docker 0.6 support.
The newest docker does not provide cli support for accessing stdout and
stderr via the logs.  The http api does provide access, so this switches
to use it for logs and a few other cases.  I'd like to put together a
PHP docker client, and that could be coming in the near future and then
could be used to replace all of these usecases.

Using stream contexts rather than curl for simplicity for now.